requirements.txt
flask
pillow
opencv-python-headless
gunicorn
moviepy
numpy